requirements.txt
# Core dependencies
torch>=1.9.0
torchdiffeq>=0.2.0
numpy>=1.21.0
scipy>=1.7.0

# Data handling
scanpy>=1.8.0
pandas>=1.3.0
anndata>=0.8.0

# Machine learning
scikit-learn>=1.0.0
inferelator>=0.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
jupyter>=1.0.0

# Development
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0